Changing long to int type

+++ main/java/py/com/progweb/prueba/ejb/ClienteDAO.java
@@ -1,19 +1,10 @@
 package py.com.progweb.prueba.ejb;
-
-import py.com.progweb.prueba.model.Agenda;
-import py.com.progweb.prueba.model.Persona;
 import py.com.progweb.prueba.model.Cliente;
-
-
 import javax.ejb.Stateless;
-import javax.inject.Inject;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
-import javax.persistence.Query;
 import java.util.List;
-import javax.persistence.EntityManager;
 import javax.persistence.TypedQuery;
-import java.util.List;
 
 @Stateless
 public class ClienteDAO {
@@ -35,7 +26,7 @@ public class ClienteDAO {
         em.remove(cliente);
     }
 
-    public Cliente obtener(Long id) {
+    public Cliente obtener(int id) {
         return em.find(Cliente.class, id);
     }
